requirements.txt
Babel==2.5.3
certifi==2018.4.16
chardet==3.0.4
idna==2.6
Pillow==5.1.0
pygame==1.9.3
pytz==2018.4
requests==2.18.4
tzlocal=1.5.1
urllib3==1.22
